Move ArrayClass.java to testclasses
